git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@357 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/security-util.c
@@ -2341,7 +2341,7 @@ check_security(
 
     myuid = getuid();
     if ((pwptr = getpwuid(myuid)) == NULL)
-        error(_("error [getpwuid(%d) fails]"), myuid);
+        error(_("error [getpwuid(%d) fails]"), (int)myuid);
 
     auth_debug(1, _("bsd security: remote host %s user %s local user %s\n"),
 		   remotehost, remoteuser, pwptr->pw_name);
